fix: symlink not supported on Windows
On the other hand, the expected eval won't get this error either.

For some reason, eval will only set $! and not $@ when symlink() is
used, so both would need to be tested after the eval block.

Also added File::Spec in order to make the paths more portable between
OSes.